--- app/src/main/java/woowacourse/shopping/cart/CartAdapter.kt
@@ -1,29 +1,0 @@
-package woowacourse.shopping.cart
-
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.cart.viewHolder.CartViewHolder
-import woowacourse.shopping.model.ProductUIModel
-
-class CartAdapter(
-    private val cartItems: List<CartItem>,
-    private val onItemClick: (ProductUIModel) -> Unit,
-    private val onItemRemove: (Int) -> Unit
-) : RecyclerView.Adapter<CartViewHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CartViewHolder {
-        return CartViewHolder.from(parent, onItemClick, onItemRemove)
-    }
-
-    override fun getItemCount(): Int {
-        return cartItems.size
-    }
-
-    override fun onBindViewHolder(holder: CartViewHolder, position: Int) {
-        holder.bind(cartItems[position])
-    }
-
-    override fun getItemViewType(position: Int): Int {
-        return position
-    }
-}
