+++ app/src/main/java/woowacourse/shopping/data/ProductFakeRepository.kt
@@ -11,7 +11,7 @@ object ProductFakeRepository : ProductRepository {
             it,
             "[사미헌] 갈비탕$it",
             12000,
-            "https://img-cf.kurly.com/cdn-cgi/image/quality=85,width=676/shop/data/goods/1648206780555l0.jpeg"
+            "https://img-cf.kurly.com/cdn-cgi/image/quality=85,width=676/shop/data/goods/1648206780555l0.jpeg",
         )
     }
