--- app/src/main/java/woowacourse/shopping/recentProduct/RecentProductItem.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.recentProduct
-
-import woowacourse.shopping.model.ProductUIModel
-
-class RecentProductItem(val product: ProductUIModel)
