--- app/src/main/java/woowacourse/shopping/recentProduct/RecentProductsAdapter.kt
@@ -1,29 +1,0 @@
-package woowacourse.shopping.recentProduct
-
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.model.ProductUIModel
-import woowacourse.shopping.recentProduct.viewHolder.RecentProductViewHolder
-import woowacourse.shopping.shopping.ProductsItemType
-
-class RecentProductsAdapter(
-    private var recentProducts: List<RecentProductItem>,
-    private val onClickItem: (data: ProductUIModel) -> Unit
-) : RecyclerView.Adapter<RecentProductViewHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecentProductViewHolder {
-        return RecentProductViewHolder.from(parent) {
-            onClickItem(recentProducts[it].product)
-        }
-    }
-
-    override fun getItemCount(): Int = recentProducts.size
-
-    override fun onBindViewHolder(holder: RecentProductViewHolder, position: Int) {
-        holder.bind(recentProducts[position])
-    }
-
-    override fun getItemViewType(position: Int): Int {
-        return ProductsItemType.TYPE_ITEM
-    }
-}
