+++ app/src/main/java/woowacourse/shopping/ui/cart/contract/CartContract.kt
@@ -1,0 +1,25 @@
+package woowacourse.shopping.ui.cart.contract
+
+import woowacourse.shopping.model.CartUIModel
+import woowacourse.shopping.model.ProductUIModel
+import woowacourse.shopping.ui.cart.CartItem
+
+interface CartContract {
+
+    interface View {
+        fun setCarts(products: List<CartItem>, cartUIModel: CartUIModel)
+        fun navigateToItemDetail(product: ProductUIModel)
+    }
+
+    interface Presenter {
+        fun setUpCarts()
+        fun pageUp()
+        fun pageDown()
+        fun removeItem(id: Int)
+        fun navigateToItemDetail(product: ProductUIModel)
+
+        fun saveOffsetState(outState: MutableMap<String, Int>)
+
+        fun restoreOffsetState(state: Map<String, Int>)
+    }
+}
