+++ app/src/main/java/woowacourse/shopping/ui/shopping/viewHolder/ItemViewHolder.kt
@@ -1,0 +1,6 @@
+package woowacourse.shopping.ui.shopping.viewHolder
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+sealed class ItemViewHolder(view: View) : RecyclerView.ViewHolder(view)
